Resolve some TODOs in pytype.
* The TODO in errors.py to not allow a filename of None is no longer necessary
  after the fix to use `filename or ""` when sorting errors.
* file_utils.get_versioned_path is used outside of pytype/pytd now, so it does
  not make sense to move it to pytd_utils.
* Remove a TODO referencing a no-longer-existing name in imports_map_loader.
* Remove TODOs to review/improve pytd files that have been in active use long
  enough that I'm pretty sure they're reasonable.
* Move some py2-only tests whose corresponding py3 bugs have been fixed to be
  target independent.
* Remove what looks to be an already-fulfilled TODO in py3/test_annotations.
* Our py2 vs. py3 test bucketization is long done, so complete a related TODO.
* Remove a TODO from cfg.py, which is basically a reference file at this point,
  since the real code has been ported to C++.

I also fixed lint errors as I encountered them.

#Squeegee

PiperOrigin-RevId: 270959139
