+++ pytype/pytd/stdlib/2/dummy_thread.pytd
@@ -1,5 +1,3 @@
-# TODO(pludemann): review this -- it was generated by:
-#         pytype --api python_runtime/v2_7/Lib/dummy_thread.py
 # This module does not exist in typeshed.
 from typing import Callable
