+++ pytype/pytd/stdlib/2and3/encodings.pytd
@@ -1,5 +1,3 @@
-# TODO(pludemann): review this -- it was generated by:
-#          pytype --api python_runtime/v2_7/Lib/encodings/__init__.py
 # This module does not exist in typeshed.
 #
 from typing import Text
