+++ pytype/tests/test_methods.py
@@ -999,5 +999,28 @@ class MethodsTest(test_base.TargetIndependentTest):
         def __enter__(self: _TFoo) -> _TFoo: ...
     """)
 
+  def testAttributeInInheritedNew(self):
+    ty = self.Infer("""
+      class Foo(object):
+        def __new__(cls, name):
+          self = super(Foo, cls).__new__(cls)
+          self.name = name
+          return self
+      class Bar(Foo):
+        def __new__(cls):
+          return super(Bar, cls).__new__(cls, "")
+    """)
+    self.assertTypesMatchPytd(ty, """
+      from typing import Any, Type, TypeVar
+      _TFoo = TypeVar("_TFoo", bound=Foo)
+      _TBar = TypeVar("_TBar", bound=Bar)
+      class Foo(object):
+        name = ...  # type: Any
+        def __new__(cls: Type[_TFoo], name) -> _TFoo
+      class Bar(Foo):
+        name = ...  # type: str
+        def __new__(cls: Type[_TBar]) -> _TBar
+    """)
+
 
 test_base.main(globals(), __name__ == "__main__")
